chore(inputs.gnmi): Cleanup code
